rotating triangle

+++ src/application.hpp
@@ -43,7 +43,7 @@ class Application{
     SwapChain swapchain{device.get_surface(), device, WIDTH, HEIGHT};
     Pipeline pipeline{device, "./shaders/simple-2d-shader.vert.spv", "./shaders/constants-shader.frag.spv", {WIDTH,HEIGHT}, &swapchain.get_image_format(), swapchain.get_depth_format()};
 
-    ObjectManager object_manager{device, swapchain.get_image_count(), Vertex::get_default_triangle()};
+    ObjectManager object_manager{device, static_cast<u_int32_t>(swapchain.get_image_count()), Vertex::get_default_triangle()};
 };
 
 }
